handle reaching the last page
